+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## Next Release
 
-### Internal 
+### Internal
 
 - Save search query to overdue search history when search query is valid
 - Add layout for overdue search screen
 - Handle user interactions for overdue search results
+- Bump Sentry to v6.1.2
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-20-8300